enterprise_system/requirements.txt
Django>=4.2,<7.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.26.0
pyotp>=2.9.0
qrcode[pil]>=7.4.2
django-axes>=6.1.0
python-decouple>=3.8
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
cryptography>=41.0.0
Pillow>=10.0.0
requests>=2.31.0
django-cors-headers>=4.3.0

# Production dependencies
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0

